minor bug fixes and renaming
