+++ src/components/nav/NavBar.js
@@ -1,5 +1,5 @@
 import { Link, useNavigate } from "react-router-dom"
-import /* addNewOrder */ "../../services/orderService.js"
+// import { addNewOrder } from "../../services/orderService.js"
 import "./NavBar.css"
 
 export const NavBar = ({ currentUser }) => {
