+++ src/views/ApplicationViews.js
@@ -4,7 +4,7 @@ import { NavBar } from "../components/nav/NavBar.js"
 import { OrdersList } from "../components/orders/OrdersList.js"
 import { CreateOrder } from "../components/orders/CreateOrder.js"
 import { EmployeesList } from "../components/employees/EmployeesList.js"
-import { SalesList } from "../components/sales/SalesView.js"
+import { SalesList } from "../components/sales/SalesList.js"
 
 export const ApplicationViews = () => {
 	const [currentUser, setCurrentUser] = useState({})
